+++ src/DustPhotonShootJobMarket.hpp
@@ -47,13 +47,13 @@ private:
   DustPhotonShootJob *_jobs[MAX_NUM_THREADS];
 
   /*! @brief Number of threads used in the calculation. */
-  const int _worksize;
+  const int_fast32_t _worksize;
 
   /*! @brief Total number of photons to propagate through the grid. */
-  unsigned int _numphoton;
+  uint_fast64_t _numphoton;
 
   /*! @brief Number of photons to shoot during a single DustPhotonShootJob. */
-  const unsigned int _jobsize;
+  const uint_fast64_t _jobsize;
 
   /*! @brief Lock used to ensure safe access to the internal photon number
    *  counters. */
@@ -77,13 +77,16 @@ public:
    */
   inline DustPhotonShootJobMarket(PhotonSource &photon_source,
                                   const DustScattering &dust_scattering,
-                                  int random_seed, DensityGrid &density_grid,
-                                  unsigned int numphoton, const CCDImage &image,
-                                  unsigned int jobsize, int worksize)
+                                  int_fast32_t random_seed,
+                                  DensityGrid &density_grid,
+                                  uint_fast64_t numphoton,
+                                  const CCDImage &image, uint_fast64_t jobsize,
+                                  int_fast32_t worksize)
       : _worksize(worksize), _numphoton(numphoton), _jobsize(jobsize) {
+
     // create a separate RandomGenerator for each thread.
     // create a single PhotonShootJob for each thread.
-    for (int i = 0; i < _worksize; ++i) {
+    for (int_fast32_t i = 0; i < _worksize; ++i) {
       _jobs[i] = new DustPhotonShootJob(photon_source, dust_scattering,
                                         random_seed + i, density_grid, image);
     }
@@ -95,7 +98,7 @@ public:
    * Deletes the internal job array.
    */
   inline ~DustPhotonShootJobMarket() {
-    for (int i = 0; i < _worksize; ++i) {
+    for (int_fast32_t i = 0; i < _worksize; ++i) {
       delete _jobs[i];
     }
   }
@@ -106,7 +109,7 @@ public:
    *
    * @param worksize Number of parallel threads that will be used.
    */
-  inline void set_worksize(int worksize) {}
+  inline void set_worksize(int_fast32_t worksize) {}
 
   /**
    * @brief Set the number of photons.
@@ -116,7 +119,7 @@ public:
    *
    * @param numphoton New number of photons.
    */
-  inline void set_numphoton(unsigned int numphoton) { _numphoton = numphoton; }
+  inline void set_numphoton(uint_fast64_t numphoton) { _numphoton = numphoton; }
 
   /**
    * @brief Update the given CCDImage.
@@ -124,7 +127,7 @@ public:
    * @param image CCDImage to update.
    */
   inline void update_image(CCDImage &image) {
-    for (int i = 0; i < _worksize; ++i) {
+    for (int_fast32_t i = 0; i < _worksize; ++i) {
       _jobs[i]->update_image(image);
     }
   }
@@ -136,8 +139,8 @@ public:
    * context).
    * @return DustPhotonShootJob.
    */
-  inline DustPhotonShootJob *get_job(int thread_id) {
-    unsigned int jobsize = std::max(_numphoton / (10 * _worksize), _jobsize);
+  inline DustPhotonShootJob *get_job(int_fast32_t thread_id) {
+    uint_fast64_t jobsize = std::max(_numphoton / (10 * _worksize), _jobsize);
     _lock.lock();
     if (jobsize >= _numphoton) {
       jobsize = _numphoton;
